HACK HACK HACK - rename midonet metadata interface to midometa
This hack renames the midolman metadata interface to midometa
to prevent a conflict of variable names inside puppet.  One of
the puppet modules is creating $network_* variables to correspond
to each of the configured network interfaces and therefore creates
a $network_metadata variable.  There are 13 other puppet manifests
that try to create a variable called $network_metadata and fail
because of this.  In the puppet-midonet module, there are another
5-10 more uses of this variable name.

This hack is based on a quick binary edit of one of the class files
in midolman.jar to change the hard-coded interface name to midometa.